Creation of predata and data impuation of missing data.

prep_predata(
  segdata,
  gridfile_name,
  varenviro,
  do_log_enviro,
  varphysio,
  do_log_physio,
  imputation = "Amelia",
  shape,
  layer,
  saturate_predata = F,
  saturate_segdata = F,
  inbox_poly = T
)

Arguments

segdata
"Segdata" data.frame, built with prepare_data_effort.

gridfile_name
Grid of the study area with all the environmental variables which allows to build predata.

varenviro
Vector containing all dynamic environmental variables (e.g. chlorophyl, SST).

do_log_enviro
Vector precising which variables among varenviro that are necessary to transform to neperian logarithm + 1 (log1p).

varphysio
Vector containing non-dynamic environmental variables (i.e depth, dist200m).

do_log_physio
Vector precising which variables among varphysio that are necessary to transform to neperian logarithm + 1 (log1p).

imputation
Data imputation method to missing values of segdata and predata. 2 methods allowed :

"PCA" : PCA method from missMDA.

"Amelia" : Amelia method from Amelia.

Default method is PCA.

shape
Shapefile of the study area. It can be either a SpatialPolygonsDataFrame class object, in this case it is not necessary to give shape_layer argument. Or it can be the name of the shape object with its extension ".shp" (ex : "data/studyAreaShapefile.shp").

saturate_predata
Boolean. If TRUE, saturate function is applied on all varenviro an varphysio columns of segdata. Saturate function excludes extreme valuesand keep values between quantiles 95

saturate_segdataBoolean. If TRUE, saturate function is applied on all varenviro an varphysio columns of predata. Saturate function excludes extreme valuesand keep values between quantiles 95 This function return a list containing :

predata : data.frame of predata.

segdata : data.frame of segdata.

pca_pred : Output of PCA function on segdata.

pca_seg : Output of PCA function on predata.

Creation of predata, a data.frame containing environmental parameters in a grid of the study area. Also it allows to impute missing data for predata and segdata.