Création de predata et imputation de données manquantes

Cette fonction permet de créer predata, le data.frame contenant les données environnementales dans la grille de la zone d'étude. Cette fonction permet également l'imputation de données manquantes pour les data.frames segdata et predata

prep_predata(
  segdata,
  gridfile_name,
  varenviro,
  do_log_enviro,
  varphysio,
  do_log_physio,
  imputation = "Amelia",
  shape,
  layer,
  saturate_predata = F,
  saturate_segdata = F,
  inbox_poly = T
)

Arguments

segdata
"Segdata" data.frame, built with prepare_data_effort.

gridfile_name
Grille de la zone d'étude avec toutes les variables d'intérêt qui permettront la fabrication de predata.

varenviro
Vecteur contenant les variables environnementales dynamiques (i.e chlorophylles, SST).

do_log_enviro
vecteur précisant parmi les variables environnementales (varenviro), lesquelles sont à transformer en logarithme néperien + 1 (log1p).

varphysio
Vecteur contenant les variables environnementales non-dynamiques (i.e depth, dist200m).

do_log_physio
vecteur précisant parmi les variables physiologiques (varphysio), lesquelles sont à transformer en logarithme néperien + 1 (log1p).

imputation
Méthode permettant l'imputation de données manquantes pour segdata et predata. Parmi les méthodes il y a :

"PCA" : par méthode ACP. méthode issue du package missMDA.

"Amelia" : par méthode Amelia. méthode issue du pakage Amelia.

Par défaut la méthode est "PCA".

shape
shapefile délimitant la zone où la prédiction est effectuée.

layer
Layer du shpaefile donnant les limite de la zone de prédiction.

proj_grid
Systême de projection du fichier de gride.

Value

Cette fonction renvoit une liste contenant :

predata : Un data.frame contenant predata

segdata : Un data.frame contenant segdata

pca_pred : Une sortie de la fonction PCA sur le data.frame segdata

pca_seg : Une sortie de la fonction PCA sur le data.frame predata

Examples